reset

+++ reset_password.php
@@ -26,38 +26,64 @@ echo "<div class='col-sm-12'>";
     
 // get given a Id
 $id=isset($_GET['Id']) ? $_GET['Id'] : die("Id not found.");
+
  
 // pass Id to the user table
 $user->Id = $id;
-  
+   $passwordErr = "";
     // if form was posted
         if($_POST){
-                   
-            // set values to object properties
-            $user->password=$_POST['new_password'];
-        
-            // reset password
-            if($user->updatePassword()){
-                echo "<div class='alert alert-info'>Password was reset. Please <a href='{$home_url}login.php'>login.</a></div>";
+            // read the password
+            $id_exists = $user->idExists();
+            
+            $passwordLength = strlen($_POST['new_password']);
+
+            // validate login
+            if ($id_exists && password_verify($_POST['password'], $user->password) && $passwordLength > 6 ){
+                            // set values to object properties
+                    $user->password=$_POST['new_password'];
+                
+                    // reset password
+                    if($user->updatePassword()){
+                        echo "<div class='alert alert-info'>Password was reset. Please <a href='{$home_url}login.php'>login.</a></div>";
+                    }
+                
+                    else{
+                        echo "<div class='alert alert-danger'>Unable to reset password.</div>";
+                    }
+
+            }else{
+               
+                echo "<div class='alert alert-danger'>incorrect current password or your new password is less than 6 </div>";
             }
-        
-            else{
-                echo "<div class='alert alert-danger'>Unable to reset password.</div>";
-            }
+            
+            
         }
         
         echo "<form action='" . htmlspecialchars($_SERVER["PHP_SELF"]) . "?Id={$id}' method='post'>
-            <table class='table table-hover table-responsive table-bordered'>
+        <p style=\"color:red;\"> $passwordErr </p>
+
+        <table class='table table-hover table-responsive table-bordered'>
+            
+            
+                <tr>
+                <td>current Password</td>
+                <td><input type='password' name='password' class='form-control' required></td>
+                </tr>
+            
                 
                 <tr>
                     <td>New Password</td>
                     <td><input type='password' name='new_password' class='form-control' required></td>
                 </tr>
+                
                 <tr>
                     <td></td>
                     <td><button type='submit' class='btn btn-primary'>Reset</button></td>
                 </tr>
+
             </table>
+            
         </form>";
